Fold the five copied exclusion skip blocks into the base, derive glob layouts from the declaration
Review follow-ups from bpintea.

The five spec suites each carried the same SKIPPED_TESTS field, shouldSkipTest override and
contains-then-find double lookup, differing only in the suite token. That block now lives once in
AbstractExternalSourceSpecTestCase behind exclusionSuiteToken(). The default is the codec-stripped
format, which is correct for csv, tsv, ndjson, orc and parquet; parquet-rs (whose format is
"parquet") and ndjson-compressed override it, because for those the default resolves to another
suite's exclusion set -- parquet-rs would have taken parquet's single entry instead of its own 45.
Both NDJSON suites previously discarded the declared reason and logged a hard-coded
SchemaAdaptingIterator message that names a defect no surviving entry has; every suite now reports
the reason the declaration states.

The two HTTP glob-skips still decided "is this a glob layout?" by substring on _multifile and
_hive -- the last copy of the layout convention this PR set out to centralise, and one that silently
fails to skip a layout named anything else. Both now ask the declaration. That removes the need for
all eight suffix constants, five of which had no readers left, and with them the precedence
reasoning for _hive_shadow: every glob layout matches on its own suffix.

Also fixes the uint32 list arm in ParquetFixtureGenerator, which appended longValue() into a
physical INT32 element while the scalar arm truncates with intValue() and documents why.

Verified: executed-set diff against the pre-change run shows 0 removed; all five suites keep their
exact skip counts, parquet-rs included at 483/587.

+++ x-pack/plugin/esql/qa/server/src/parquetFixtureGenerator/java/org/elasticsearch/xpack/esql/datasources/ParquetFixtureGenerator.java
@@ -409,7 +409,9 @@ public final class ParquetFixtureGenerator {
             case "version" -> throw new IllegalArgumentException(
                 "declared [version] in a list: Parquet has no version type, same as the scalar arm"
             );
-            case "uint32" -> listElement.append("element", ((Number) elem).longValue());
+            // Same reason as the scalar arm: the physical list element is INT32, so truncating the
+            // uint32 long to int preserves the raw bit pattern the column stores.
+            case "uint32" -> listElement.append("element", ((Number) elem).intValue());
             case "uint16" -> listElement.append("element", ((Number) elem).intValue());
             case "uint64" -> listElement.append("element", ((Number) elem).longValue());
             default -> listElement.append("element", elem.toString());
